+++ examples/web/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @workglow/web
 
+## 0.0.101
+
+### Patch Changes
+
+- Promote task config to first class schema, remove old name prop in favor of title
+- Updated dependencies
+  - @workglow/task-graph@0.0.101
+  - @workglow/tasks@0.0.101
+  - @workglow/test@0.0.101
+  - @workglow/ai@0.0.101
+  - @workglow/ai-provider@0.0.101
+  - @workglow/debug@0.0.101
+  - @workglow/job-queue@0.0.101
+  - @workglow/sqlite@0.0.101
+  - @workglow/storage@0.0.101
+  - @workglow/util@0.0.101
+
 ## 0.0.100
 
 ### Patch Changes