+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @workglow/ai
 
+## 0.0.101
+
+### Patch Changes
+
+- Promote task config to first class schema, remove old name prop in favor of title
+- Updated dependencies
+  - @workglow/task-graph@0.0.101
+  - @workglow/dataset@0.0.101
+  - @workglow/job-queue@0.0.101
+  - @workglow/storage@0.0.101
+  - @workglow/util@0.0.101
+
 ## 0.0.100
 
 ### Patch Changes